added error handling for reviews:  edit and post new
